# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Reservas - Universidad",
        "description": "Backend para la gestión de reservas de recursos (salas, equipos, etc.) con autenticación JWT y roles admin/usuario.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservas/{id}/historial": {
            "get": {
                "tags": [
                    "Historial de Reservas"
                ],
                "summary": "Ver historial de una reserva",
                "description": "Obtiene el historial completo de cambios de una reserva específica, ordenado cronológicamente. Los usuarios normales solo pueden ver el historial de sus propias reservas.",
                "operationId": "8404ced2049f604fb2fb720680cc423a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial de la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "reserva_id": {
                                                "type": "integer",
                                                "example": 15
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "accion": {
                                                "description": "Tipo de acción: creada, actualizada_admin, actualizada_usuario, cancelada_admin, cancelada_usuario",
                                                "type": "string",
                                                "example": "creada"
                                            },
                                            "detalle": {
                                                "type": "string",
                                                "example": "Reserva creada para el recurso Aula 101 del 2025-12-10 08:00:00 al 2025-12-10 10:00:00."
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-12-07T10:30:00.000000Z"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "juan@universidad.edu"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para ver el historial de esta reserva"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}